requirements.txt
google-cloud-bigquery==2.13.1
pyarrow==3.0.0
google-cloud-storage==1.37.1
pandas==1.2.3
plotly==4.14.3
psutil==5.8.0
kaleido==0.2.1
anytree==2.8.0
markdownify==0.9.0
slack-sdk==3.7.0
python-pptx==0.6.19
xhtml2pdf==0.2.5